requirements.txt
# Core
openai
numpy
faiss-cpu
prompt_toolkit
rich
llama-index-core
llama-index-embeddings-openai
llama-index-retrievers-bm25

# HF / training / evaluation
torch
transformers
datasets
accelerate
peft
trl
bitsandbytes
unsloth

# Utilities
huggingface_hub[cli]
